// src/enums/ActionCode.ts
export enum ActionCode {
  MAIN_LANGUAGE = 'mlng',
  MAIN_SETTINGS = 'mstg',
  MAIN_BOTS = 'mbts',
  MAIN_SOURCES = 'msrc',
  MAIN_ADD_CHANNEL = 'machl',
  MAIN_HELP = 'mhlp',
  MAIN_CONTACT = 'mctc',
  MAIN_ADMIN = 'madm',
  SETTINGS_LANGUAGE = 'slng',
  LANGUAGE_SELECT = 'lsct',
  BOTS_ADD = 'badd',
  BOTS_SELECT = 'bsct',
  BOT_DELETE = 'bdel',
  BOT_CHANNELS = 'bchnl',
  SOURCES_CHANNEL_SELECT = 'scsct',
  SOURCES_LIST = 'slst',
  SOURCES_LIST_ADD_RSS = 'sars',
  SOURCES_LIST_SELECT = 'ssct',
  SOURCE_DELETE = 'sdel',
  SOURCE_CHECK = 'schk',
  CHANNELS_SELECT = 'csct',
  CHANNEL_DELETE = 'cdel',
  ADMIN_STAT = 'asta',
}